+++ pandas/tests/extension/base/methods.py
@@ -193,14 +193,15 @@ class BaseMethodsTests(BaseExtensionTests):
             expected = pd.Series(
                 orig_data1._from_sequence(
                     [a + b for (a, b) in zip(list(orig_data1), list(orig_data2))]
-                )
+                ),
+                dtype=orig_data1.dtype,
             )
         self.assert_series_equal(result, expected)
-
         val = s1.iloc[0]
         result = s1.combine(val, lambda x1, x2: x1 + x2)
         expected = pd.Series(
-            orig_data1._from_sequence([a + val for a in list(orig_data1)])
+            orig_data1._from_sequence([a + val for a in list(orig_data1)]),
+            dtype=orig_data1.dtype,
         )
         self.assert_series_equal(result, expected)
